+++ apps/api/.dockerignore
