+++ apps/api/scripts/README.md
