+++ apps/api/scripts/upload-openapi-spec.ts
