+++ apps/api/src/routes/course/katex.ts
