+++ apps/api/src/routes/course/lesson.ts
