+++ apps/api/src/routes/mail.ts
