+++ apps/api/src/types/course/lesson.ts
