+++ apps/api/src/types/mail.ts
