+++ apps/api/src/utils/auth/validate-user.ts
