+++ apps/api/src/utils/certificate.ts
