+++ apps/api/src/utils/course.ts
