+++ apps/api/src/utils/mail.ts
