+++ apps/api/src/utils/openapi/index.ts
