+++ apps/dashboard/.env.example
@@ -14,7 +14,7 @@ UNSPLASH_API_KEY=
 OPENAI_API_KEY=
 
 # Sever URL - Optional if you want to use video upload and pdf downloads
-# This is the deployment URL of apps/backend
+# This is the deployment URL of apps/api
 PUBLIC_SERVER_URL=
 
 # For getting metadata for signup users.
