+++ apps/dashboard/package.json
@@ -21,7 +21,7 @@
   },
   "dependencies": {
     "@carbon/charts-svelte": "^1.13.6",
-    "@cio/backend": "workspace:*",
+    "@cio/api": "workspace:*",
     "@lemonsqueezy/lemonsqueezy.js": "^2.2.0",
     "@polar-sh/sveltekit": "^0.2.1",
     "@supabase/supabase-js": "^2.31.0",
