+++ apps/dashboard/src/lib/utils/functions/tinymce/plugins.ts
@@ -80,7 +80,7 @@ export const addMathPlugin = (tinymce: any) => {
         onSubmit: async function (w) {
           let content = output.exportAs('html');
 
-          // The generated content always points to https://latex.codecogs.com/svg.image, this means we have to always rely on their server to serve formulas. We have a dedicated endpoint for generating formulas on our backend service: GET `/katex`. We will need to manually update the URL in the generate HTML to point to our server:
+          // The generated content always points to https://latex.codecogs.com/svg.image, this means we have to always rely on their server to serve formulas. We have a dedicated endpoint for generating formulas on our api service: GET `/katex`. We will need to manually update the URL in the generate HTML to point to our server:
           // https://latex.codecogs.com/svg.image
 
           // 1. Get only the URL in the img tag generated by codecogs library
