+++ apps/docs/pages/contributor-guides/cloudflare-setup.mdx
@@ -4,4 +4,4 @@ import { Callout } from 'nextra/components';
 
 # Cloudflare Setup
 
-In this guide, you will learn how to get your cloudflare environment variables. We use Cloudflare R2 for video uploads and the aim of this guide is to help you get the right variables to properly setup our backend.
+In this guide, you will learn how to get your cloudflare environment variables. We use Cloudflare R2 for video uploads and the aim of this guide is to help you get the right variables to properly setup our api.
